rename js folder to client
