+++ app/client/ui/components/footer.react.js
